azure_live_scraper/requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
azure-storage-blob>=12.14.0
pandas>=1.5.0
urllib3>=1.26.0 